Api and server file restructuring
